Put all ObjectAccessControls request in same file.
One of the ObjectAccessControls request classes was initially in its own
file, but we are now keeping them all together.
