+++ google/cloud/storage/storage_client.bzl
@@ -29,7 +29,6 @@ storage_client_HDRS = [
     "internal/google_application_default_credentials_file.h",
     "internal/http_response.h",
     "internal/insert_object_media_request.h",
-    "internal/list_object_acl_request.h",
     "internal/list_objects_request.h",
     "internal/logging_client.h",
     "internal/metadata_parser.h",
@@ -80,7 +79,6 @@ storage_client_SRCS = [
     "internal/get_object_metadata_request.cc",
     "internal/google_application_default_credentials_file.cc",
     "internal/insert_object_media_request.cc",
-    "internal/list_object_acl_request.cc",
     "internal/list_objects_request.cc",
     "internal/logging_client.cc",
     "internal/metadata_parser.cc",
