+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -17,7 +17,6 @@ storage_client_unit_tests = [
     "internal/get_object_metadata_request_test.cc",
     "internal/google_application_default_credentials_file_test.cc",
     "internal/insert_object_media_request_test.cc",
-    "internal/list_object_acl_request_test.cc",
     "internal/list_objects_request_test.cc",
     "internal/logging_client_test.cc",
     "internal/metadata_parser_test.cc",
